# youki original config