--- a/crates/libcontainer/src/config.rs
+++ b/crates/libcontainer/src/config.rs
@@ -13,6 +13,7 @@ use crate::utils;
 /// A configuration for passing information obtained during container creation to other commands.
 /// Keeping the information to a minimum improves performance.
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
+#[non_exhaustive]
 pub struct YoukiConfig {
     pub hooks: Option<Hooks>,
     pub cgroup_path: PathBuf,