--- a/crates/libcontainer/src/container/container_start.rs
+++ b/crates/libcontainer/src/container/container_start.rs
@@ -1,4 +1,5 @@
 use crate::{
+    config::YoukiConfig,
     hooks,
     notify_socket::{NotifySocket, NOTIFY_FILE},
 };
@@ -39,10 +40,9 @@ impl Container {
             bail!(err_msg);
         }
 
-        let spec = self
-            .spec()
+        let config = YoukiConfig::load(self.root.join("config.json"))
             .with_context(|| format!("failed to load runtime spec for container {}", self.id()))?;
-        if let Some(hooks) = spec.hooks() {
+        if let Some(hooks) = config.hooks.as_ref() {
             // While prestart is marked as deprecated in the OCI spec, the docker and integration test still
             // uses it.
             #[allow(deprecated)]
@@ -60,7 +60,7 @@ impl Container {
 
         // Run post start hooks. It runs after the container process is started.
         // It is called in the runtime namespace.
-        if let Some(hooks) = spec.hooks() {
+        if let Some(hooks) = config.hooks.as_ref() {
             hooks::run_hooks(hooks.poststart().as_ref(), Some(self))
                 .with_context(|| "failed to run post start hooks")?;
         }