--- a/crates/libcontainer/src/container/init_builder.rs
+++ b/crates/libcontainer/src/container/init_builder.rs
@@ -48,7 +48,7 @@ impl<'a> InitContainerBuilder<'a> {
             .set_annotations(spec.annotations().clone());
 
         let config = YoukiConfig::from_spec(&spec, container.id())?;
-        config.save(&container_dir.join("yconfig.json"))?;
+        config.save(&container_dir)?;
 
         unistd::chdir(&container_dir)?;
         let notify_path = container_dir.join(NOTIFY_FILE);