@yihuaf 2021-11-01
I am not sure I follow the argument. Can you elaborate on why the spec "is too much information and isn't good for performance"?