@utam0k 2021-11-01
> I am not sure I follow the argument. Can you elaborate on why the spec "is too much information and isn't good for performance"?

@yihuaf 
Sure. A `start` and `delete` commands need the information about cgrop_path and hooks via the file(it is actually `config.json`)
Now, this file is the file copied spec file. There is much unnecessary information(e.g. devices, seccomp...) In the spec file when the start or delete command executes.
This PR prevents copying a large file (`Spec`) by only copying with necessary information when executing the `start` or `delete` commands when executing the create command.