@utam0k 2021-11-01
@yihuaf I don't think it's a very big improvement either, but the current form is the one I skimmed over in its initial form (almost got the sample working!). The current form is one that I slacked off on in its initial form (I slacked off because I was so close to getting the sample working), so fixing that is a big part of it.