@utam0k 2021-11-03
@yihuaf @Furisto 
I tried to measure the performance. What do you think of this? I was surprised that it was faster than I expected. I guess it's probably due to the reduced JSON parsing part.
```console
$ $ hyperfine --prepare 'sudo sync; echo 3 | sudo tee /proc/sys/vm/drop_caches' --warmup 10 --min-runs 100 'sudo ./youki create -b tutorial a && sudo ./youki start a && sudo ./youki delete -f a'
```
Results
|Target| Time (mean ± σ)|   Range (min … max)|
|:---:|:---:|:---:|
|[This PR](d68cf57b8bf4cb87ccbd8c2e114edcd02ce51dda)(Release)| 183.2 ms ±  52.5 ms  | 111.4 ms … 517.5 ms  |
|[Main Branch](e11c3f31e4af2009f0b609a20972e66846bdf77c)(Release)| 227.2 ms ±  45.0 ms | 64.2 ms … 972.6 ms |