@Furisto 2021-11-03
Huh, proofs again that you should always benchmark your performance improvements. I would not have expected such a big difference. Strange that the variance between runs is quite considerable.