@YJDoc2 2021-11-15
Hey, this looks good :+1:
Can you add @yihuaf 's suggestion :
> We may also want to put a comment down that the change is for performance reason. When in the future this struct becomes closer to the size of spec, we should switch back.

Maybe add this in the doc-draft.md as well.

Also, do you think that the config.rs should be renamed to youki_config.rs or something, to avoid misunderstanding it as general runtime config or so?