@utam0k 2021-11-18
@YJDoc2 
Thanks for your comments :>

I commented on this area. I think the doc-draft is a bit much, but what do you think?
https://github.com/containers/youki/pull/447/files#diff-10c989310fea87b855588bfc33a18d95b4779d090fa372aecb10c01465be7c8aR15-R16
> Hey, this looks good +1 Can you add @yihuaf 's suggestion :
> 
> > We may also want to put a comment down that the change is for performance reason. When in the future this struct becomes closer to the size of spec, we should switch back.
> 
> Maybe add this in the doc-draft.md as well.

I renamed.
> Also, do you think that the config.rs should be renamed to youki_config.rs or something, to avoid misunderstanding it as general runtime config or so?